Implementation 2 30KB 0.5MB

+++ include/layer1_hal.h
@@ -15,13 +15,13 @@
 
 #include "config.h"
 
-/* --- Display Flush Buffer Structure --- */
+/* --- Display Flush Buffer Structure (LVGL Compatible) --- */
 typedef struct {
     int16_t x1;
     int16_t y1;
     int16_t x2;
     int16_t y2;
-    const uint32_t *pixel_color_p;
+    const uint8_t *pixel_color_p;
 } display_area_t;
 
 /* --- Function Declarations --- */
@@ -38,8 +38,8 @@ void layer1_queue_touch(int16_t x, int16_t y, bool pressed);
 bool layer1_poll_button_event(input_key_t *key_out);
 bool layer1_poll_touch_event(int16_t *x, int16_t *y, bool *pressed);
 
-/* Display Flush Callback (called by rendering engine) */
-void layer1_display_flush_cb(const display_area_t *area, const uint32_t *color_p);
+/* Display Flush Callback (called by LVGL rendering engine) */
+void layer1_display_flush_cb(const display_area_t *area, const uint8_t *color_p);
 
 /* Heartbeat reporter from hardware side */
 void layer1_report_hardware_alive(void);
